arch: rename DropdownAction to ListGeoAction
