+++ shared/territory/listGeo.contract.ts
